+++ CHANGELOG.md
@@ -31,3 +31,4 @@ replaced with `parseBody`, `parseUploadedFiles`, and `parseRawRequestBuffer`, re
 as well as `query`.
 * Removed `Angel.injections` and `RequestContext.injections`.
 * Removed `Angel.inject`.
+* Removed a dependency on `package:pool`, which also meant removing `AngelHttp.throttle`.